promising

+++ src/vbo.hpp
@@ -12,7 +12,7 @@ struct VBO : Buffer {
 
     static inline struct Quad : File { Quad() : File("quad.obj", "o quad\n\nv -1 -1 0\nv 1 -1 0\nv -1 1 0\nv 1 1 0\n\nvt 0 1\nvt 1 1\nvt 0 0\nvt 1 0 \n\nf 1/1 2/2 3/3 \nf 2/2 3/3 4/4") { } } quad;
 
-    static inline Struct *vertice = &Struct::create("Vertice").add<glm::vec2>("POSITION").range(-1,1,0).add<glm::vec2>("UV").add<glm::vec2>("NORMALIZED").add<int>("OBJ").add<int>("LAYER");
+    static inline Struct *vertice = &Struct::create("Vertice").add<glm::vec2>("POSITION").range(-1,1,0).add<glm::vec2>("UV").add<glm::vec2>("NORMALIZED").add<float>("OBJ").add<float>("LAYER");
 
     static inline Struct *indice = &Struct::create("Indice").add<int>("vert1").add<int>( "vert2").add<int>("vert3");
 
@@ -40,7 +40,7 @@ struct VBO : Buffer {
 
     void draw(int count = 1);
 
-    void addQuad(float w, float h, float x, float y, float id = 0);
+    void addQuad(float w, float h, float x, float y, int id = 0);
 
     bool add(File *file = &quad, int id = 0);
     bool add_noupload(File *file = &quad, int id = 0);
